Started Implementation of a Poll Manager

+++ src/main/java/ampullen/registration/RegistrationListener.java
@@ -106,7 +106,7 @@ public class RegistrationListener extends ListenerAdapter {
                 }
             case Phone:
                 try {
-                    Integer.parseInt(message);
+                    Long.parseLong(message);
                     if(message.length() > 9) {
                         ultimateLinzUser.UserData.Phone = message;
                         return true;
